Kestrel: widen RX data_rate to 16-bit + two-radio HE monitor probe
RxAtrib.data_rate was uint8_t, truncating the 9-bit AX/VHT datarate code:
every HE (0x180+) and VHT2SS+ (0x110+) rate collapsed onto an HT/legacy code,
so a Kestrel used as an HE monitor could not even report an HE rate. Widened to
uint16_t (the two read-consumers are legacy-range compares, unaffected; the
Jaguar2/3 assignments that silently truncated VHT now report the true code).

Adds tests/kestrel_two_radio_he.sh: one Kestrel airs HE-rate beacons, the other
monitors — a two-AX-radio HE TX->RX probe (the 11ac 8812AU witness can't decode
HE). It currently surfaces a real gap: the 8832CU monitor decodes only a few
ambient frames and none of the co-located beacons, with an ADDCK S0/S1 timeout
at init — its RX ADC-DC calibration (8852b-derived DACK/ADDCK) does not complete
on the 8852C, degrading sensitivity (~40x fewer ambient frames than the 8852BU,
which DACKs clean). The vendor ships a distinct halrf_dack_8852c; porting it is
task #11. Comment on the CCA gate corrected to current state (DACK/RX-DCK are
applied; carrier-sense TX still needs IQK/DPK).

+++ src/kestrel/RtlKestrelDevice.cpp
@@ -284,7 +284,7 @@ void RtlKestrelDevice::StartRxLoop(Action_ParsedRadioPacket packetProcessor) {
             p.RxAtrib.pkt_len = static_cast<uint16_t>(f.payload_len);
             p.RxAtrib.crc_err = f.crc_err;
             p.RxAtrib.icv_err = f.icv_err;
-            p.RxAtrib.data_rate = static_cast<uint8_t>(f.rx_rate);
+            p.RxAtrib.data_rate = f.rx_rate; /* 9-bit AX code (HE >= 0x180) */
             p.RxAtrib.tsfl = f.freerun_cnt;
             p.RxAtrib.rssi[0] = _last_rssi[0];
             p.RxAtrib.rssi[1] = _last_rssi[1];
